Move non-container casting util out of ContainerUtils
