+++ src/ast/utility/NodeMapper.h
@@ -17,6 +17,7 @@
 #pragma once
 
 #include "souffle/utility/ContainerUtil.h"
+#include "souffle/utility/MiscUtil.h"
 #include <cassert>
 #include <memory>
