+++ src/ram/AbstractConditional.h
@@ -19,7 +19,7 @@
 #include "ram/Node.h"
 #include "ram/NodeMapper.h"
 #include "ram/Operation.h"
-#include "souffle/utility/ContainerUtil.h"
+#include "souffle/utility/MiscUtil.h"
 #include <cassert>
 #include <memory>
 #include <string>
