+++ src/ram/NestedOperation.h
@@ -17,7 +17,7 @@
 #include "ram/Node.h"
 #include "ram/NodeMapper.h"
 #include "ram/Operation.h"
-#include "souffle/utility/ContainerUtil.h"
+#include "souffle/utility/MiscUtil.h"
 #include <cassert>
 #include <iosfwd>
 #include <memory>
